refactoring reduce operations
